pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "note_organizer"
version = "0.1.0"
description = "AI-powered note organization microservice"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "sentence-transformers==2.2.2",
    "openai==1.6.0",
    "python-dotenv==1.0.0",
    "markdown==3.4.3",
    "nltk==3.8.1",
    "tqdm==4.66.1",
    "pyyaml==6.0.1",
    "numpy==1.24.3",
    "scikit-learn==1.3.0",
    "fastapi==0.104.0",
    "uvicorn==0.23.2",
    "pydantic==2.4.2",
    "watchfiles==0.21.0",
    "dspy-ai==2.3.5",
    "sqlmodel==0.0.8",
    "loguru==0.7.2",
    "matplotlib==3.7.3",
    "wordcloud==1.9.2",
    "cce>=0.1.2",
    "torch>=1.10.0",
    "httpx>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black==23.10.0",
    "isort==5.12.0",
    "mypy==1.6.1",
    "pytest==7.4.2",
    "pytest-cov==4.1.0",
]
dashboard = [
    "streamlit>=1.20.0",
    "plotly>=5.10.0",
]

[project.scripts]
note-organizer = "note_organizer.cli.main:main"
note-organizer-api = "note_organizer.api.server:start"
note-organizer-dashboard = "note_organizer.dashboard.app:main [dashboard]"

[tool.setuptools]
packages = ["note_organizer"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 